# backend/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: KubeLens API
  version: v0.4.0
  description: Core API contract for KubeLens dashboard and automation clients.
servers:
  - url: /api
paths:
...
  /pods:
    get:
      operationId: listPods
      summary: List pods across namespaces
      responses:
        "200":
          description: Pod list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PodSummary"
    post:
      operationId: createPod
      summary: Create pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PodCreateRequest"
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pods/{namespace}/{name}:
    get:
      operationId: getPodDetail
      summary: Get pod detail
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pod detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodDetail"
    delete:
      operationId: deletePod
      summary: Delete pod
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /pods/{namespace}/{name}/events:
    get:
      operationId: getPodEvents
      summary: Fetch pod events
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pod events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/K8sEvent"
  /pods/{namespace}/{name}/logs:
    get:
      operationId: getPodLogs
      summary: Fetch pod logs
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: lines
          required: false
          schema:
            type: integer
            minimum: 1
            default: 50
        - in: query
          name: container
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Pod log output
          content:
            text/plain:
              schema:
                type: string
  /pods/{namespace}/{name}/logs/stream:
    get:
      operationId: streamPodLogs
      summary: Stream pod logs
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: tailLines
          required: false
          schema:
            type: integer
            minimum: 1
            default: 100
        - in: query
          name: container
          required: false
          schema:
            type: string
        - in: query
          name: follow
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Streaming log output
          content:
            text/event-stream:
              schema:
                type: string
  /pods/{namespace}/{name}/describe:
    get:
      operationId: getPodDescribe
      summary: Get kubectl describe output
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Describe output
          content:
            text/plain:
              schema:
                type: string
  /pods/{namespace}/{name}/restart:
    post:
      operationId: restartPod
      summary: Restart pod
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
  /nodes/{name}/pods:
    get:
      operationId: listNodePods
      summary: List pods scheduled on a node
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pod list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PodSummary"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PodSummary:
      type: object
      required: [id, name, namespace, status, cpu, memory, age, restarts]
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        nodeName:
          type: string
        status:
          type: string
        cpu:
          type: string
        memory:
          type: string
        age:
          type: string
        restarts:
          type: integer
    PodDetail:
      allOf:
        - $ref: "#/components/schemas/PodSummary"
        - type: object
          properties:
            nodeName:
              type: string
            hostIP:
              type: string
            podIP:
              type: string
            containers:
              type: array
              items:
                $ref: "#/components/schemas/ContainerSpec"
            volumes:
              type: array
              items:
                $ref: "#/components/schemas/NamedVolume"
    ContainerSpec:
      type: object
      required: [name]
      properties:
        name:
          type: string
        image:
          type: string
        env:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
        volumeMounts:
          type: array
          items:
            $ref: "#/components/schemas/VolumeMount"
        resources:
          $ref: "#/components/schemas/ContainerResources"
    EnvVar:
      type: object
      required: [name]
      properties:
        name:
          type: string
        value:
          type: string
    VolumeMount:
      type: object
      properties:
        name:
          type: string
        mountPath:
          type: string
    ContainerResources:
      type: object
      properties:
        requests:
          $ref: "#/components/schemas/ResourcePairs"
        limits:
          $ref: "#/components/schemas/ResourcePairs"
    ResourcePairs:
      type: object
      properties:
        cpu:
          type: string
        memory:
          type: string
    NamedVolume:
      type: object
      properties:
        name:
          type: string
...
    K8sEvent:
      type: object
      properties:
        type:
          type: string
        reason:
          type: string
        age:
          type: string
        from:
          type: string
        message:
          type: string
        namespace:
          type: string
        resource:
          type: string
        resourceKind:
          type: string
        count:
          type: integer
        lastTimestamp:
          type: string
...
    PodCreateRequest:
      type: object
      required: [namespace, name, image]
      properties:
        namespace:
          type: string
        name:
          type: string
        image:
          type: string
...
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...